requirements.txt
torch>=2.0.0
transformers>=4.30.0
lightgbm>=4.0.0
biopython
scikit-learn>=1.3.0
pandas
numpy
tqdm
h5py
filelock
requests
matplotlib
seaborn
colorama
umap-learn
